Create array and receive the player's input
Takes the initial play from the human player and immediately generates cpu move. At the end of the program, if there is no winner, the program will continue to ask for additional moves from the player. Trying to choose a position already in play results in being asked to choose again.